kernel: use u8 to represent #allow #upcall
Since we are storing the number of all kernel managemed allows and
upcalls in a single u32, each field only gets a few bits. To enforce
this by the type system we change the fields to u8.

However, because of the current implementation of IPC, which is related
to the humber of possible processes on the board, this gets a little
invasive and complicated.